♻️ Move Pf2eSkillBonus into Pf2eJsonNodeReader
